feat: add new logic that course connect to subject, and subject connect to subjectTopic; refactor code to this logic

+++ src/main/java/edu/tinkoff/tinkoffbackendacademypetproject/dto/requests/SubjectByCourseNumberRequestDTO.java
@@ -10,6 +10,6 @@ import lombok.EqualsAndHashCode;
 public class SubjectByCourseNumberRequestDTO extends PageRequestDto {
     @Min(1)
     @Schema(description = "Номер курса предмета", example = "1")
-    private Long course;
+    private Integer course;
 }
